+++ profiles/preview/policyinsights/mgmt/policyinsights/policyinsightsapi/models.go
@@ -19,11 +19,8 @@
 
 package policyinsightsapi
 
-import original "github.com/Azure/azure-sdk-for-go/services/policyinsights/mgmt/2019-10-01/policyinsights/policyinsightsapi"
+import original "github.com/Azure/azure-sdk-for-go/services/policyinsights/mgmt/2018-04-04/policyinsights/policyinsightsapi"
 
 type OperationsClientAPI = original.OperationsClientAPI
 type PolicyEventsClientAPI = original.PolicyEventsClientAPI
-type PolicyMetadataClientAPI = original.PolicyMetadataClientAPI
 type PolicyStatesClientAPI = original.PolicyStatesClientAPI
-type PolicyTrackedResourcesClientAPI = original.PolicyTrackedResourcesClientAPI
-type RemediationsClientAPI = original.RemediationsClientAPI
